+++ fixed_list.h
@@ -8,6 +8,7 @@
 
 namespace rde
 {
+
 template<typename T, size_t TCapacity>
 class fixed_list
 {
@@ -125,7 +126,7 @@ private:
 
 public:
 	typedef T												value_type;
-	typedef int												size_type;
+	typedef size_t											size_type;
 	typedef node_iterator<node*, T*, T&>					iterator;
 	typedef node_iterator<const node*, const T*, const T&>	const_iterator;
